:recycle: [#4577] Factor out generic bits of ReactSelect usage
